LayoutTests/platform/ios-sim-deprecated/http/tests/security/postMessage/origin-unaffected-by-document-domain-expected.txt

window.location.href = http://127.0.0.1:8000/security/postMessage/origin-unaffected-by-document-domain.html
document.domain = 0.0.1
waiting...
Received message: data="Received message: data="Message from parent" origin="http://127.0.0.1:8000"" origin="http://localhost:8000"
Received message: data="Received message: data="done" origin="http://127.0.0.1:8000"" origin="http://localhost:8000"
